Crate rusoto_mq

Amazon MQ is a managed message broker service for Apache ActiveMQ that makes it easy to set up and operate message brokers in the cloud. A message broker allows software applications and components to communicate using various programming languages, operating systems, and formal messaging protocols.

If you're using the service, you're probably looking for MQClient and MQ.

Structs

AvailabilityZone
Name of the availability zone.

BrokerEngineType
Types of broker engines.

BrokerInstance
Returns information about all brokers.

BrokerInstanceOption
Option for host instance type.

BrokerSummary
The Amazon Resource Name (ARN) of the broker.

Configuration
Returns information about all configurations.

ConfigurationId
A list of information about the configuration.

ConfigurationRevision
Returns information about the specified configuration revision.

Configurations
Broker configuration information

CreateBrokerRequest
Creates a broker using the specified properties.

CreateBrokerResponse

CreateConfigurationRequest
Creates a new configuration for the specified configuration name. Amazon MQ uses the default configuration (the engine type and version).

CreateConfigurationResponse

CreateTagsRequest
A map of the key-value pairs for the resource tag.

CreateUserRequest
Creates a new ActiveMQ user.

CreateUserResponse

DeleteBrokerRequest

DeleteBrokerResponse

DeleteTagsRequest

DeleteUserRequest

DeleteUserResponse

DescribeBrokerEngineTypesRequest

DescribeBrokerEngineTypesResponse

DescribeBrokerInstanceOptionsRequest

DescribeBrokerInstanceOptionsResponse

DescribeBrokerRequest

DescribeBrokerResponse

DescribeConfigurationRequest

DescribeConfigurationResponse

DescribeConfigurationRevisionRequest

DescribeConfigurationRevisionResponse

DescribeUserRequest

DescribeUserResponse

EngineVersion
Id of the engine version.

ListBrokersRequest

ListBrokersResponse

ListConfigurationRevisionsRequest

ListConfigurationRevisionsResponse

ListConfigurationsRequest

ListConfigurationsResponse

ListTagsRequest

ListTagsResponse

ListUsersRequest

ListUsersResponse

Logs
The list of information about logs to be enabled for the specified broker.

LogsSummary
The list of information about logs currently enabled and pending to be deployed for the specified broker.

MQClient
A client for the AmazonMQ API.

PendingLogs
The list of information about logs to be enabled for the specified broker.

RebootBrokerRequest

RebootBrokerResponse

SanitizationWarning
Returns information about the XML element or attribute that was sanitized in the configuration.

UpdateBrokerRequest
Updates the broker using the specified properties.

UpdateBrokerResponse

UpdateConfigurationRequest
Updates the specified configuration.

UpdateConfigurationResponse

UpdateUserRequest
Updates the information for an ActiveMQ user.

UpdateUserResponse

User
An ActiveMQ user associated with the broker.

UserPendingChanges
Returns information about the status of the changes pending for the ActiveMQ user.

UserSummary
Returns a list of all ActiveMQ users.

WeeklyStartTime
The scheduled time period relative to UTC during which Amazon MQ begins to apply pending updates or patches to the broker.

Enums

CreateBrokerError
Errors returned by CreateBroker

CreateConfigurationError
Errors returned by CreateConfiguration

CreateTagsError
Errors returned by CreateTags

CreateUserError
Errors returned by CreateUser

DeleteBrokerError
Errors returned by DeleteBroker

DeleteTagsError
Errors returned by DeleteTags

DeleteUserError
Errors returned by DeleteUser

DescribeBrokerEngineTypesError
Errors returned by DescribeBrokerEngineTypes

DescribeBrokerError
Errors returned by DescribeBroker

DescribeBrokerInstanceOptionsError
Errors returned by DescribeBrokerInstanceOptions

DescribeConfigurationError
Errors returned by DescribeConfiguration

DescribeConfigurationRevisionError
Errors returned by DescribeConfigurationRevision

DescribeUserError
Errors returned by DescribeUser

ListBrokersError
Errors returned by ListBrokers

ListConfigurationRevisionsError
Errors returned by ListConfigurationRevisions

ListConfigurationsError
Errors returned by ListConfigurations

ListTagsError
Errors returned by ListTags

ListUsersError
Errors returned by ListUsers

RebootBrokerError
Errors returned by RebootBroker

UpdateBrokerError
Errors returned by UpdateBroker

UpdateConfigurationError
Errors returned by UpdateConfiguration

UpdateUserError
Errors returned by UpdateUser

Traits

MQ
Trait representing the capabilities of the AmazonMQ API. AmazonMQ clients implement this trait.