Struct rusoto_mq::DescribeBrokerOutput

pub struct DescribeBrokerOutput {
    pub auto_minor_version_upgrade: Option<bool>,
    pub broker_arn: Option<String>,
    pub broker_id: Option<String>,
    pub broker_instances: Option<Vec<BrokerInstance>>,
    pub broker_name: Option<String>,
    pub broker_state: Option<String>,
    pub configurations: Option<Configurations>,
    pub created: Option<f64>,
    pub deployment_mode: Option<String>,
    pub engine_type: Option<String>,
    pub engine_version: Option<String>,
    pub host_instance_type: Option<String>,
    pub logs: Option<LogsSummary>,
    pub maintenance_window_start_time: Option<WeeklyStartTime>,
    pub publicly_accessible: Option<bool>,
    pub security_groups: Option<Vec<String>>,
    pub subnet_ids: Option<Vec<String>>,
    pub users: Option<Vec<UserSummary>>,
}

The version of the broker engine. Note: Currently, Amazon MQ supports only 5.15.0.

Fields
auto_minor_version_upgrade: Option<bool>

Required. Enables automatic upgrades to new minor versions for brokers, as Apache releases the versions. The automatic upgrades occur during the maintenance window of the broker or after a manual broker reboot.

broker_arn: Option<String>

The Amazon Resource Name (ARN) of the broker.

broker_id: Option<String>

The unique ID that Amazon MQ generates for the broker.

broker_instances: Option<Vec<BrokerInstance>>

A list of information about allocated brokers.

broker_name: Option<String>

The name of the broker. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain whitespaces, brackets, wildcard characters, or special characters.

broker_state: Option<String>

The status of the broker.

configurations: Option<Configurations>

The list of all revisions for the specified configuration.

created: Option<f64>

The time when the broker was created.

deployment_mode: Option<String>

Required. The deployment mode of the broker.

engine_type: Option<String>

Required. The type of broker engine. Note: Currently, Amazon MQ supports only ACTIVEMQ.

engine_version: Option<String>

The version of the broker engine. Note: Currently, Amazon MQ supports only 5.15.0.

host_instance_type: Option<String>

The broker's instance type.

logs: Option<LogsSummary>

The list of information about logs currently enabled and pending to be deployed for the specified broker.

maintenance_window_start_time: Option<WeeklyStartTime>

The parameters that determine the WeeklyStartTime.

publicly_accessible: Option<bool>

Required. Enables connections from applications outside of the VPC that hosts the broker's subnets.

security_groups: Option<Vec<String>>

Required. The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

subnet_ids: Option<Vec<String>>

The list of groups (2 maximum) that define which subnets and IP ranges the broker can use from different Availability Zones. A SINGLEINSTANCE deployment requires one subnet (for example, the default subnet). An ACTIVESTANDBYMULTIAZ deployment requires two subnets.

users: Option<Vec<UserSummary>>

The list of all ActiveMQ usernames for the specified broker.