Trait rusoto_mq::MQ

pub trait MQ {
    fn create_broker(
        &self, 
        input: CreateBrokerRequest
    ) -> RusotoFuture<CreateBrokerResponse, CreateBrokerError>;

    fn create_configuration(
        &self, 
        input: CreateConfigurationRequest
    ) -> RusotoFuture<CreateConfigurationResponse, CreateConfigurationError>;

    fn create_tags(
        &self, 
        input: CreateTagsRequest
    ) -> RusotoFuture<(), CreateTagsError>;

    fn create_user(
        &self, 
        input: CreateUserRequest
    ) -> RusotoFuture<CreateUserResponse, CreateUserError>;

    fn delete_broker(
        &self, 
        input: DeleteBrokerRequest
    ) -> RusotoFuture<DeleteBrokerResponse, DeleteBrokerError>;

    fn delete_tags(
        &self, 
        input: DeleteTagsRequest
    ) -> RusotoFuture<(), DeleteTagsError>;

    fn delete_user(
        &self, 
        input: DeleteUserRequest
    ) -> RusotoFuture<DeleteUserResponse, DeleteUserError>;

    fn describe_broker(
        &self, 
        input: DescribeBrokerRequest
    ) -> RusotoFuture<DescribeBrokerResponse, DescribeBrokerError>;

    fn describe_broker_engine_types(
        &self, 
        input: DescribeBrokerEngineTypesRequest
    ) -> RusotoFuture<DescribeBrokerEngineTypesResponse, DescribeBrokerEngineTypesError>;

    fn describe_broker_instance_options(
        &self, 
        input: DescribeBrokerInstanceOptionsRequest
    ) -> RusotoFuture<DescribeBrokerInstanceOptionsResponse, DescribeBrokerInstanceOptionsError>;

    fn describe_configuration(
        &self, 
        input: DescribeConfigurationRequest
    ) -> RusotoFuture<DescribeConfigurationResponse, DescribeConfigurationError>;

    fn describe_configuration_revision(
        &self, 
        input: DescribeConfigurationRevisionRequest
    ) -> RusotoFuture<DescribeConfigurationRevisionResponse, DescribeConfigurationRevisionError>;

    fn describe_user(
        &self, 
        input: DescribeUserRequest
    ) -> RusotoFuture<DescribeUserResponse, DescribeUserError>;

    fn list_brokers(
        &self, 
        input: ListBrokersRequest
    ) -> RusotoFuture<ListBrokersResponse, ListBrokersError>;

    fn list_configuration_revisions(
        &self, 
        input: ListConfigurationRevisionsRequest
    ) -> RusotoFuture<ListConfigurationRevisionsResponse, ListConfigurationRevisionsError>;

    fn list_configurations(
        &self, 
        input: ListConfigurationsRequest
    ) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>;

    fn list_tags(
        &self, 
        input: ListTagsRequest
    ) -> RusotoFuture<ListTagsResponse, ListTagsError>;

    fn list_users(
        &self, 
        input: ListUsersRequest
    ) -> RusotoFuture<ListUsersResponse, ListUsersError>;

    fn reboot_broker(
        &self, 
        input: RebootBrokerRequest
    ) -> RusotoFuture<RebootBrokerResponse, RebootBrokerError>;

    fn update_broker(
        &self, 
        input: UpdateBrokerRequest
    ) -> RusotoFuture<UpdateBrokerResponse, UpdateBrokerError>;

    fn update_configuration(
        &self, 
        input: UpdateConfigurationRequest
    ) -> RusotoFuture<UpdateConfigurationResponse, UpdateConfigurationError>;

    fn update_user(
        &self, 
        input: UpdateUserRequest
    ) -> RusotoFuture<UpdateUserResponse, UpdateUserError>;
}

Trait representing the capabilities of the AmazonMQ API. AmazonMQ clients implement this trait.

Required methods

fn create_broker(
&self,
input: CreateBrokerRequest
) -> RusotoFuture<CreateBrokerResponse, CreateBrokerError>

Creates a broker. Note: This API is asynchronous.

fn create_configuration(
&self,
input: CreateConfigurationRequest
) -> RusotoFuture<CreateConfigurationResponse, CreateConfigurationError>

Creates a new configuration for the specified configuration name. Amazon MQ uses the default configuration (the engine type and version).

fn create_tags(
&self,
input: CreateTagsRequest
) -> RusotoFuture<(), CreateTagsError>

Add a tag to a resource.

fn create_user(
&self,
input: CreateUserRequest
) -> RusotoFuture<CreateUserResponse, CreateUserError>

Creates an ActiveMQ user.

fn delete_broker(
&self,
input: DeleteBrokerRequest
) -> RusotoFuture<DeleteBrokerResponse, DeleteBrokerError>

Deletes a broker. Note: This API is asynchronous.

fn delete_tags(
&self,
input: DeleteTagsRequest
) -> RusotoFuture<(), DeleteTagsError>

Removes a tag from a resource.

fn delete_user(
&self,
input: DeleteUserRequest
) -> RusotoFuture<DeleteUserResponse, DeleteUserError>

Deletes an ActiveMQ user.

fn describe_broker(
&self,
input: DescribeBrokerRequest
) -> RusotoFuture<DescribeBrokerResponse, DescribeBrokerError>

Returns information about the specified broker.

fn describe_broker_engine_types(
&self,
input: DescribeBrokerEngineTypesRequest
) -> RusotoFuture<DescribeBrokerEngineTypesResponse, DescribeBrokerEngineTypesError>

Describe available engine types and versions.

fn describe_broker_instance_options(
&self,
input: DescribeBrokerInstanceOptionsRequest
) -> RusotoFuture<DescribeBrokerInstanceOptionsResponse, DescribeBrokerInstanceOptionsError>

Describe available broker instance options.

fn describe_configuration(
&self,
input: DescribeConfigurationRequest
) -> RusotoFuture<DescribeConfigurationResponse, DescribeConfigurationError>

Returns information about the specified configuration.

fn describe_configuration_revision(
&self,
input: DescribeConfigurationRevisionRequest
) -> RusotoFuture<DescribeConfigurationRevisionResponse, DescribeConfigurationRevisionError>

Returns the specified configuration revision for the specified configuration.

fn describe_user(
&self,
input: DescribeUserRequest
) -> RusotoFuture<DescribeUserResponse, DescribeUserError>

Returns information about an ActiveMQ user.

fn list_brokers(
&self,
input: ListBrokersRequest
) -> RusotoFuture<ListBrokersResponse, ListBrokersError>

Returns a list of all brokers.

fn list_configuration_revisions(
&self,
input: ListConfigurationRevisionsRequest
) -> RusotoFuture<ListConfigurationRevisionsResponse, ListConfigurationRevisionsError>

Returns a list of all revisions for the specified configuration.

fn list_configurations(
&self,
input: ListConfigurationsRequest
) -> RusotoFuture<ListConfigurationsResponse, ListConfigurationsError>

Returns a list of all configurations.

fn list_tags(
&self,
input: ListTagsRequest
) -> RusotoFuture<ListTagsResponse, ListTagsError>

Lists tags for a resource.

fn list_users(
&self,
input: ListUsersRequest
) -> RusotoFuture<ListUsersResponse, ListUsersError>

Returns a list of all ActiveMQ users.

fn reboot_broker(
&self,
input: RebootBrokerRequest
) -> RusotoFuture<RebootBrokerResponse, RebootBrokerError>

Reboots a broker. Note: This API is asynchronous.

fn update_broker(
&self,
input: UpdateBrokerRequest
) -> RusotoFuture<UpdateBrokerResponse, UpdateBrokerError>

Adds a pending configuration change to a broker.

fn update_configuration(
&self,
input: UpdateConfigurationRequest
) -> RusotoFuture<UpdateConfigurationResponse, UpdateConfigurationError>

Updates the specified configuration.

fn update_user(
&self,
input: UpdateUserRequest
) -> RusotoFuture<UpdateUserResponse, UpdateUserError>

Updates the information for an ActiveMQ user.

Loading content...