Crate rusoto_mq

Amazon MQ is a managed message broker service for Apache ActiveMQ that makes it easy to set up and operate message brokers in the cloud. A message broker allows software applications and components to communicate using various programming languages, operating systems, and formal messaging protocols.

If you're using the service, you're probably looking for MQClient and MQ.

Structs

BrokerInstance
Returns information about all brokers.

BrokerSummary
The Amazon Resource Name (ARN) of the broker.

Configuration
Returns information about all configurations.

ConfigurationId
A list of information about the configuration.

ConfigurationRevision
Returns information about the specified configuration revision.

Configurations
Broker configuration information

CreateBrokerInput
Required. The time period during which Amazon MQ applies pending updates or patches to the broker.

CreateBrokerOutput
Returns information about the created broker.

CreateBrokerRequest
Creates a broker using the specified properties.

CreateBrokerResponse

CreateConfigurationInput
Creates a new configuration for the specified configuration name. Amazon MQ uses the default configuration (the engine type and version).

CreateConfigurationOutput
Returns information about the created configuration.

CreateConfigurationRequest
Creates a new configuration for the specified configuration name. Amazon MQ uses the default configuration (the engine type and version).

CreateConfigurationResponse

CreateUserInput
Creates a new ActiveMQ user.

CreateUserRequest
Creates a new ActiveMQ user.

CreateUserResponse

DeleteBrokerOutput
Returns information about the deleted broker.

DeleteBrokerRequest

DeleteBrokerResponse

DeleteUserRequest

DeleteUserResponse

DescribeBrokerOutput
The version of the broker engine. Note: Currently, Amazon MQ supports only 5.15.0.

DescribeBrokerRequest

DescribeBrokerResponse

DescribeConfigurationRequest

DescribeConfigurationResponse

DescribeConfigurationRevisionOutput
Returns the specified configuration revision for the specified configuration.

DescribeConfigurationRevisionRequest

DescribeConfigurationRevisionResponse

DescribeUserOutput
Returns information about an ActiveMQ user.

DescribeUserRequest

DescribeUserResponse

ListBrokersOutput
A list of information about all brokers.

ListBrokersRequest

ListBrokersResponse

ListConfigurationRevisionsOutput
Returns a list of all revisions for the specified configuration.

ListConfigurationRevisionsRequest

ListConfigurationRevisionsResponse

ListConfigurationsOutput
Returns a list of all configurations.

ListConfigurationsRequest

ListConfigurationsResponse

ListUsersOutput
Returns a list of all ActiveMQ users.

ListUsersRequest

ListUsersResponse

Logs
The list of information about logs to be enabled for the specified broker.

LogsSummary
The list of information about logs currently enabled and pending to be deployed for the specified broker.

MQClient
A client for the AmazonMQ API.

MQError
Returns information about an error.

PendingLogs
The list of information about logs to be enabled for the specified broker.

RebootBrokerRequest

RebootBrokerResponse

SanitizationWarning
Returns information about the XML element or attribute that was sanitized in the configuration.

UpdateBrokerInput
Updates the broker using the specified properties.

UpdateBrokerOutput
Returns information about the updated broker.

UpdateBrokerRequest
Updates the broker using the specified properties.

UpdateBrokerResponse

UpdateConfigurationInput
Updates the specified configuration.

UpdateConfigurationOutput
Returns information about the updated configuration.

UpdateConfigurationRequest
Updates the specified configuration.

UpdateConfigurationResponse

UpdateUserInput
Updates the information for an ActiveMQ user.

UpdateUserRequest
Updates the information for an ActiveMQ user.

UpdateUserResponse

User
An ActiveMQ user associated with the broker.

UserPendingChanges
Returns information about the status of the changes pending for the ActiveMQ user.

UserSummary
Returns a list of all ActiveMQ users.

WeeklyStartTime
The scheduled time period relative to UTC during which Amazon MQ begins to apply pending updates or patches to the broker.

Enums

CreateBrokerError
Errors returned by CreateBroker

CreateConfigurationError
Errors returned by CreateConfiguration

CreateUserError
Errors returned by CreateUser

DeleteBrokerError
Errors returned by DeleteBroker

DeleteUserError
Errors returned by DeleteUser

DescribeBrokerError
Errors returned by DescribeBroker

DescribeConfigurationError
Errors returned by DescribeConfiguration

DescribeConfigurationRevisionError
Errors returned by DescribeConfigurationRevision

DescribeUserError
Errors returned by DescribeUser

ListBrokersError
Errors returned by ListBrokers

ListConfigurationRevisionsError
Errors returned by ListConfigurationRevisions

ListConfigurationsError
Errors returned by ListConfigurations

ListUsersError
Errors returned by ListUsers

RebootBrokerError
Errors returned by RebootBroker

UpdateBrokerError
Errors returned by UpdateBroker

UpdateConfigurationError
Errors returned by UpdateConfiguration

UpdateUserError
Errors returned by UpdateUser

Traits

MQ
Trait representing the capabilities of the AmazonMQ API. AmazonMQ clients implement this trait.