Struct rusoto_mq::MQClient

pub struct MQClient { /* fields omitted */ }

A client for the AmazonMQ API.

Methods

impl MQClient

pub fn new(region: Region) -> MQClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> MQClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,